snapshot

+++ bin/git-commit.sh
@@ -1,6 +1,6 @@
 clear
 git status
-git add $M2/doc/*.ods
+#git add $M2/doc/*.ods
 git add $M2/java/MildredCacheMonitor/*.xml
 git add $M2/java/MildredCacheMonitor/*.sh
 git add $M2/java/MildredCacheMonitor/*.xml
@@ -9,27 +9,28 @@ git add $M2/java/MildredCacheMonitor/src/main/java/com/angrysurfer/mildred/ui/*.
 git add $M2/java/MildredCacheMonitor/src/main/java/com/angrysurfer/mildred/ui/swing/*.java
 git add $M2/java/MildredCacheMonitor/src/main/java/com/angrysurfer/mildred/ui/javafx/*.java
 git add $M2/java/MildredCacheMonitor/src/main/resources/com/angrysurfer/mildred/ui/javafx/*.fxml
+
 git add $M2/python/test/*.py
-rm $M2/python/test/*.pyc
+[ ! -f "*.pyc" ] && rm $M2/python/test/*.pyc
 git add $M2/python/mildred/*.py
-rm $M2/python/mildred/*.pyc
+[ ! -f "*.pyc" ] && rm $M2/python/mildred/*.pyc
 git add $M2/python/mildred/core/*.py
-rm $M2/python/mildred/core/*.pyc
+[ ! -f "*.pyc" ] && rm $M2/python/mildred/core/*.pyc
 git add $M2/python/mildred/db/*.py
-rm $M2/python/mildred/db/*.pyc
+[ ! -f "*.pyc" ] && rm $M2/python/mildred/db/*.pyc
 git add $M2/python/mildred/db/generated/*.py
-rm $M2/python/mildred/db/generated/*.pyc
+[ ! -f "*.pyc" ] && rm $M2/python/mildred/db/generated/*.pyc
 # git add $M2/python/mildred/db/*.sh
 git add $M2/python/mildred/future/*.py
-rm $M2/python/mildred/future/*.pyc
+[ ! -f "*.pyc" ] && rm $M2/python/mildred/future/*.pyc
 git add $M2/python/mildred/desktop/*.py
-rm $M2/python/mildred/desktop/*.pyc
+[ ! -f "*.pyc" ] && rm $M2/python/mildred/desktop/*.pyc
 git add $M2/python/mildred/scratch/*.py
-rm $M2/python/mildred/scratch/*.pyc
+[ ! -f "*.pyc" ] && rm $M2/python/mildred/scratch/*.pyc
 
 git add $M2/db/bak/*.sql
 git add $M2/db/setup/*.sql
 git add $M2/bin/*.sh
-git add $M2/db/*.sql
-git add $M2/db/*.mwb
+git add $M2/db/design/*.sql
+git add $M2/db/design/*.mwb
 git status
